LayoutTests/imported/w3c/web-platform-tests/css/css-ruby/line-break-around-ruby-001-expected.txt


PASS OP+ID
PASS ID+CL
PASS span+CL
FAIL OP+ruby assert_approx_equals: expected 3 +/- 0.1 but got 2
PASS ruby+CL
FAIL OP+ruby+CL assert_approx_equals: expected 3 +/- 0.1 but got 2
あ（かき
あい）か
あす）か
あ（べるか
あべる）か
あ（べる）か
